Fix group phase randomness not working.
The intention behind the integer cast was probably to cast the whole
expression to int, but instead it cast only the first part, which has
to be float. It's not necessary to cast the whole expression, since it
will be done implicitly by the compiler when assigning to outpos, so
just remove it.

+++ src/Synth/OscilGen.cpp
@@ -1326,7 +1326,7 @@ int OscilGen::getPhase()
         return 0;
 
     int outpos;
-    outpos = (int)(prng.numRandom() * 2.0f - 1.0f) * synth->oscilsize_f * (params->Prand - 64.0f) / 64.0f;
+    outpos = (prng.numRandom() * 2.0f - 1.0f) * synth->oscilsize_f * (params->Prand - 64.0f) / 64.0f;
     outpos = (outpos + 2 * synth->oscilsize) % synth->oscilsize;
     return outpos;
 }
